refactor: the code structure
