+++ packages/web3-providers/src/NFTScan/apis/Solana.ts
@@ -1,0 +1,6 @@
+import type { ChainId, SchemaType } from "@masknet/web3-shared-solana";
+import type { NonFungibleTokenAPI } from "../../types";
+
+export class NFTScanSolanaAPI implements NonFungibleTokenAPI.Provider<ChainId, SchemaType> {
+
+}
